# Fixed #33308 -- Added support for psycopg version 3