--- a/django/contrib/postgres/search.py
+++ b/django/contrib/postgres/search.py
@@ -39,6 +39,11 @@ class SearchQueryField(Field):
         return "tsquery"
 
 
+class _Float4Field(Field):
+    def db_type(self, connection):
+        return "float4"
+
+
 class SearchConfig(Expression):
     def __init__(self, config):
         super().__init__()
@@ -138,7 +143,11 @@ class SearchVector(SearchVectorCombinable, Func):
         if clone.weight:
             weight_sql, extra_params = compiler.compile(clone.weight)
             sql = "setweight({}, {})".format(sql, weight_sql)
-        return sql, config_params + params + extra_params
+
+        # These parameters must be bound on the client side because we may
+        # want to create an index on this expression.
+        sql = connection.ops.compose_sql(sql, config_params + params + extra_params)
+        return sql, []
 
 
 class CombinedSearchVector(SearchVectorCombinable, CombinedExpression):
@@ -244,6 +253,8 @@ class SearchRank(Func):
         normalization=None,
         cover_density=False,
     ):
+        from .fields.array import ArrayField
+
         if not hasattr(vector, "resolve_expression"):
             vector = SearchVector(vector)
         if not hasattr(query, "resolve_expression"):
@@ -252,6 +263,7 @@ class SearchRank(Func):
         if weights is not None:
             if not hasattr(weights, "resolve_expression"):
                 weights = Value(weights)
+            weights = Cast(weights, ArrayField(_Float4Field()))
             expressions = (weights,) + expressions
         if normalization is not None:
             if not hasattr(normalization, "resolve_expression"):