--- a/django/db/backends/base/features.py
+++ b/django/db/backends/base/features.py
@@ -164,6 +164,8 @@ class BaseDatabaseFeatures:
     # Can we roll back DDL in a transaction?
     can_rollback_ddl = False
 
+    schema_editor_uses_clientside_param_binding = False
+
     # Does it support operations requiring references rename in a transaction?
     supports_atomic_references_rename = True
 
@@ -334,6 +336,9 @@ class BaseDatabaseFeatures:
 
     # Does the backend support the logical XOR operator?
     supports_logical_xor = False
+
+    # Set to (exception, message) if null characters in text are disallowed.
+    prohibits_null_characters_in_text_exception = None
 
     # Collation names for use by the Django test suite.
     test_collations = {