… django/db/backends/postgresql/operations.py
… django/db/backends/postgresql/psycopg_any.py
… django/db/backends/postgresql/schema.py
… django/db/models/lookups.py
… docs/conf.py
… docs/ref/contrib/gis/install/index.txt
… docs/ref/contrib/gis/install/postgis.txt
… docs/ref/contrib/postgres/fields.txt
… docs/ref/contrib/postgres/forms.txt
… docs/ref/databases.txt
… docs/releases/1.11.txt
… docs/releases/4.2.txt
… docs/topics/db/transactions.txt
… docs/topics/install.txt
… tests/backends/postgresql/tests.py
… tests/backends/tests.py
… tests/db_functions/datetime/test_extract_trunc.py
… tests/db_utils/tests.py
… tests/fixtures/tests.py
… tests/gis_tests/tests.py
… tests/model_fields/test_jsonfield.py
… tests/postgres_tests/test_apps.py
… tests/postgres_tests/test_array.py
… tests/schema/test_logging.py