What did I do? I took
https://github.com/dvarrazzo/django-psycopg3-backend and blackified it +
ported over most (all?) new commits. I am now opening this on GitHub to
be able to nicely diff and start a discussion about whether we can
support psycopg2 & 3 easiyl from the same codebase (I think we can).

----
I (i.e. @felixxm) have the following plan to move this forward:
- [x]  merge #16245,
- [x]  squash commits,
- [x]  rebase,
- [x]  revert unnecessary changes, e.g.
  - [x]  https://github.com/django/django/pull/15687#discussion_r1036784854
- [x]  merge more small cleanups:
  - #16345
  - #16346
  - #16347
  - #16351
  - #16352
  - #16353
  - #16362
  - #16363
  - #16368
  - #16371
  - 2f38f7b8f9bd65905dc6bec4f3e2a30b5da8e338
  - 1d90c9b1132482d446ca8375c151e894002c9e8d
  - db7bb3b64e469fbb5c79e7b5b2fcb890434aa60f
- [x]  push (or wait) for docs changes,
- [x]  move import changes to a separate PR:
  - 2ebfbd894e21e1656c1e1f32d98b8df7a32d3649
- [x]  make detailed review,
- [x]  push (or wait) for final edits,
- [ ]  merge :tada: