@apollo13 2022-06-05
@timgraham I can reproduce your query issue when I set `supports_order_by_nulls_modifier = False` (which it probably is on old cockroachdb versions: https://github.com/cockroachdb/django-cockroachdb/blob/master/django_cockroachdb/features.py#L60 -- did you test against an old version?). That said, the combination of `supports_order_by_nulls_modifier = False` & `supports_order_column_alias = True` certainly shows there is a problem in the new code.