@timgraham 2022-06-06
When you have a chance, could please rebase the branch to pick up the bump to Django 4.2? That will make my ongoing testing a little easier. Thanks.