@gnat 2022-08-16
Just tested this with CockroachDB on [1043b36](https://github.com/django/django/commit/1043b3655d02d277a5c200788d381d8a9ed078d9). First impressions: Seems to work! Brilliant job.

https://github.com/cockroachdb/django-cockroachdb Needs to be updated with psycopg3 @timgraham. I was able to get it running by just changing the following lines:

* In `django_cockroachdb/base.py`
```python
import psycopg
#import psycopg2  # noqa
#import psycopg2.extensions  # noqa
#import psycopg2.extras  # noqa
```
* In `django_cockroachdb/operations.py`
```python
from psycopg import errors
#from psycopg2 import errorcodes
#....
if (getattr(exc.__cause__, 'pgcode', '') != errors.SERIALIZATION_FAILURE or
#if (getattr(exc.__cause__, 'pgcode', '') != errorcodes.SERIALIZATION_FAILURE or
```
* Commented out version check in `django_cockroachdb/utils.py`

There's just too many Cockroach-specific settings to reasonably merge it into the core Postgres backend, so updating `django-cockroachdb` is essential work for Cockroach DB.

![image](https://user-images.githubusercontent.com/24665/184909616-520f48a7-4244-4559-a3dd-b3d2335f794c.png)