@timgraham 2022-08-19
@apollo13 Will you tackle the PostGIS backend? (First issue: `from psycopg2 import Binary` in `django/contrib/gis/db/backends/postgis/adapter.py`)