@thinkwelltwd 2022-11-29
> see if you get a big difference timing wise?

A partial run of 512 tests:

|main|psycopg3|speedup|
|:--------|:---------|:-------|
|190.102|184.098|3.158%|