@thinkwelltwd 2022-11-29
> using the [wait-select branch](https://github.com/psycopg/psycopg/tree/wait-select)?

|main|psycopg3|speedup|
|:--------|:---------|:-------|
|190.102|184.098|3.158%|

|main|wait-select|speedup|
|:--------|:---------|:-------|
|190.102|181.274|4.643%|

@dvarrazzo, I hafta say I'm well pleased to get almost 5% bump just by updating psycopg!! Many thanks to you for your work well done on psycopg3!!